[Fleet] Add missing step showing K85 instructions for managed enrollment
